requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
mysql-connector-python==8.2.0
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.3
pytz==2023.3.post1
requests==2.31.0
urllib3==2.1.0
